+++ lib/constants/constants.dart
@@ -1,5 +1,5 @@
 // TODO: Don't forget to change app version here
-const String appVersion = '0.9.12';
+const String appVersion = '0.9.13';
 
 const allowedNoteExtensions = ['.md', '.markdown', '.txt', '.me'];
 const allowedImageExtensions = ['.jpg', '.jpeg', '.png'];